backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Pillow>=10.3.0

# AI/ML Libraries
spacy>=3.7.2
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.1.1
scikit-learn>=1.3.2

# Document Processing
PyMuPDF>=1.23.8
python-docx>=1.1.0
pytesseract>=0.3.10

# Utilities
pandas>=2.1.3
numpy>=1.26.2
regex>=2023.10.3
django-cors-headers==4.3.1
